+++ service/src/main/java/org/grnet/status/services/ProfileService.java
@@ -8,6 +8,7 @@ import org.grnet.status.dtos.profile.aggregation.AggregationProfileResponse;
 import org.grnet.status.dtos.profile.metric.MetricProfileResponse;
 import org.grnet.status.dtos.profile.operation.OperationProfileResponse;
 import org.grnet.status.services.clients.ArgoWebApiClient;
+import org.grnet.status.services.clients.WebApiService;
 
 /**
  * Service responsible for retrieving aggregation, metric, and operation profiles from Argo Web API.
@@ -18,6 +19,9 @@ public class ProfileService {
     @Inject
     @RestClient
     ArgoWebApiClient argoWebApiClient;
+
+    @Inject
+    WebApiService webApiService;
 
     @ConfigProperty(name = "web.api.access.token")
     String accessToken;
@@ -32,6 +36,8 @@ public class ProfileService {
      */
     public AggregationProfileResponse listSpecificAggregationProfiles(String tenantId, String profileId, String date){
 
+        webApiService.validateTenantInitialized(tenantId, "Aggregation Profiles");
+
         return argoWebApiClient.listSpecificAggregationProfilesSuperAdmin(profileId, date, accessToken, tenantId);
     }
 
@@ -43,6 +49,8 @@ public class ProfileService {
      * @return aggregation profiles response
      */
     public AggregationProfileResponse listAllAggregationProfiles(String tenantId, String date){
+
+        webApiService.validateTenantInitialized(tenantId, "Aggregation Profiles");
 
         return argoWebApiClient.listAllAggregationProfilesSuperAdmin(date, accessToken, tenantId);
     }
@@ -57,6 +65,8 @@ public class ProfileService {
      */
     public MetricProfileResponse listSpecificMetricProfiles(String tenantId, String profileId, String date){
 
+        webApiService.validateTenantInitialized(tenantId, "Metric Profiles");
+
         return argoWebApiClient.listSpecificMetricProfilesSuperAdmin(profileId, date, accessToken, tenantId);
     }
 
@@ -68,6 +78,8 @@ public class ProfileService {
      * @return metric profiles response
      */
     public MetricProfileResponse listAllMetricProfiles(String tenantId, String date){
+
+        webApiService.validateTenantInitialized(tenantId, "Metric Profiles");
 
         return argoWebApiClient.listAllMetricProfilesSuperAdmin(date, accessToken, tenantId);
     }
@@ -82,6 +94,8 @@ public class ProfileService {
      */
     public OperationProfileResponse listSpecificOperationsProfiles(String tenantId, String profileId, String date){
 
+        webApiService.validateTenantInitialized(tenantId, "Operations Profiles");
+
         return argoWebApiClient.listSpecificOperationsProfilesSuperAdmin(profileId, date, accessToken, tenantId);
     }
 
@@ -93,6 +107,8 @@ public class ProfileService {
      * @return operation profiles response
      */
     public OperationProfileResponse listAllOperationsProfiles(String tenantId, String date){
+
+        webApiService.validateTenantInitialized(tenantId, "Operations Profiles");
 
         return argoWebApiClient.listAllOperationsProfilesSuperAdmin(date, accessToken, tenantId);
     }